Added unified executor and main split into files
